fix: add domain prefix * de.hatoka can be replaced if original fork will operate
