+++ .gitignore
@@ -1,8 +1,11 @@
 .idea
-target
-out
-artifacts
-nn_data.json
-src/main/java/basicneuralnetwork/Example.java
-src/main/java/basicneuralnetwork/activationfunctions/SoftmaxActivationFunction.java
-nn_data.json
+.project
+.classpath
+.settings/**
+
+target/**
+build/**
+
+out/**
+artifacts/**
+/target/
