+++ pom.xml
@@ -4,9 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.kim-marcel.basic_neural_network</groupId>
+    <groupId>de.hatoka.basic_neural_network</groupId>
     <artifactId>basic_neural_network</artifactId>
-    <version>v0.5</version>
+    <version>1.0.0-SNAPSHOT</version>
 
     <build>
         <plugins>
@@ -14,8 +14,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>17</source>
+                    <target>17</target>
                 </configuration>
             </plugin>
 
@@ -69,6 +69,19 @@
             <artifactId>gson</artifactId>
             <version>2.8.4</version>
         </dependency>
+
+		<dependency>
+		    <groupId>org.slf4j</groupId>
+		    <artifactId>slf4j-api</artifactId>
+		    <version>2.0.7</version>
+		</dependency>
+		<!-- https://mvnrepository.com/artifact/ch.qos.logback/logback-classic -->
+		<dependency>
+		    <groupId>ch.qos.logback</groupId>
+		    <artifactId>logback-classic</artifactId>
+		    <version>1.4.6</version>
+		    <scope>runtime</scope>
+		</dependency>
 
         <dependency>
             <groupId>org.junit.jupiter</groupId>
