+++ src/main/java/de/hatoka/basicneuralnetwork/WrongDimensionException.java
@@ -1,4 +1,4 @@
-package basicneuralnetwork;
+package de.hatoka.basicneuralnetwork;
 
 import java.util.Arrays;
