+++ src/main/java/de/hatoka/basicneuralnetwork/activationfunctions/ReLuActivationFunction.java
@@ -1,45 +1,43 @@
-package basicneuralnetwork.activationfunctions;
+package de.hatoka.basicneuralnetwork.activationfunctions;
 
 import org.ejml.simple.SimpleMatrix;
 
 /**
- * Created by KimFeichtinger on 20.04.18.
+ * Created by KimFeichtinger on 26.04.18.
  */
-public class SigmoidActivationFunction implements ActivationFunction {
+public class ReLuActivationFunction implements ActivationFunction {
 
-    private static final String NAME = "SIGMOID";
+    private static final String NAME = "RELU";
 
-    // Sigmoid
     public SimpleMatrix applyActivationFunctionToMatrix(SimpleMatrix input) {
         SimpleMatrix output = new SimpleMatrix(input.numRows(), input.numCols());
 
         for (int i = 0; i < input.numRows(); i++) {
             // Column is always 0 because input has only one column
             double value = input.get(i, 0);
-            double result = 1 / (1 + Math.exp(-value));
+            double result = value > 0 ? value : 0;
 
             output.set(i, 0, result);
         }
 
         // Formula:
-        // 1 / (1 + Math.exp(-input));
+        // for input < 0: 0, else input
         return output;
     }
 
-    // Derivative of Sigmoid (not real derivative because Activation function has already been applied to the input)
     public SimpleMatrix applyDerivativeOfActivationFunctionToMatrix(SimpleMatrix input) {
         SimpleMatrix output = new SimpleMatrix(input.numRows(), input.numCols());
 
         for (int i = 0; i < input.numRows(); i++) {
             // Column is always 0 because input has only one column
             double value = input.get(i, 0);
-            double result = value * (1 - value);
+            double result = value > 0 ? 1 : 0;
 
             output.set(i, 0, result);
         }
 
         // Formula:
-        // input * (1 - input);
+        // for input > 0: 1, else 0
         return output;
     }
