+++ src/test/resources/de/hatoka/basicneuralnetwork/utilities/neuro1.json
